+++ src/torchmetrics/functional/text/perplexity.py
@@ -130,10 +130,9 @@ def perplexity(preds: Tensor, target: Tensor, ignore_index: Optional[int] = None
         Perplexity value
 
     Examples:
-        >>> import torch
-        >>> gen = torch.manual_seed(42)
-        >>> preds = torch.rand(2, 8, 5, generator=gen)
-        >>> target = torch.randint(5, (2, 8), generator=gen)
+        >>> from torch import rand, randint
+        >>> preds = rand(2, 8, 5)
+        >>> target = randint(5, (2, 8))
         >>> target[0, 6:] = -100
         >>> perplexity(preds, target, ignore_index=-100)
         tensor(5.8540)