+++ src/torchmetrics/segmentation/mean_iou.py
@@ -70,20 +70,19 @@ class MeanIoU(Metric):
             If ``input_format`` is not one of ``"one-hot"`` or ``"index"``
 
     Example:
-        >>> import torch
-        >>> _ = torch.manual_seed(0)
+        >>> from torch import randint
         >>> from torchmetrics.segmentation import MeanIoU
         >>> miou = MeanIoU(num_classes=3)
-        >>> preds = torch.randint(0, 2, (10, 3, 128, 128))
-        >>> target = torch.randint(0, 2, (10, 3, 128, 128))
+        >>> preds = randint(0, 2, (10, 3, 128, 128))
+        >>> target = randint(0, 2, (10, 3, 128, 128))
         >>> miou(preds, target)
-        tensor(0.3318)
+        tensor(0.3326)
         >>> miou = MeanIoU(num_classes=3, per_class=True)
         >>> miou(preds, target)
-        tensor([0.3322, 0.3303, 0.3329])
+        tensor([0.3334, 0.3327, 0.3318])
         >>> miou = MeanIoU(num_classes=3, per_class=True, include_background=False)
         >>> miou(preds, target)
-        tensor([0.3303, 0.3329])
+        tensor([0.3327, 0.3318])
 
     """
 
@@ -111,7 +110,8 @@ class MeanIoU(Metric):
         self.input_format = input_format
 
         num_classes = num_classes - 1 if not include_background else num_classes
-        self.add_state("score", default=torch.zeros(num_classes if per_class else 1), dist_reduce_fx="mean")
+        self.add_state("score", default=torch.zeros(num_classes if per_class else 1), dist_reduce_fx="sum")
+        self.add_state("num_batches", default=torch.tensor(0), dist_reduce_fx="sum")
 
     def update(self, preds: Tensor, target: Tensor) -> None:
         """Update the state with the new data."""
@@ -120,10 +120,11 @@ class MeanIoU(Metric):
         )
         score = _mean_iou_compute(intersection, union, per_class=self.per_class)
         self.score += score.mean(0) if self.per_class else score.mean()
+        self.num_batches += 1
 
     def compute(self) -> Tensor:
-        """Update the state with the new data."""
-        return self.score  # / self.num_batches
+        """Compute the final Mean Intersection over Union (mIoU)."""
+        return self.score / self.num_batches
 
     def plot(self, val: Union[Tensor, Sequence[Tensor], None] = None, ax: Optional[_AX_TYPE] = None) -> _PLOT_OUT_TYPE:
         """Plot a single or multiple values from the metric.