+++ src/torchmetrics/text/infolm.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Any, ClassVar, Dict, List, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor
@@ -111,11 +111,24 @@ class InfoLM(Metric):
     """
 
     is_differentiable = False
-    higher_is_better = True
     preds_input_ids: List[Tensor]
     preds_attention_mask: List[Tensor]
     target_input_ids: List[Tensor]
     target_attention_mask: List[Tensor]
+
+    _information_measure_higher_is_better: ClassVar = {
+        # following values are <0
+        "kl_divergence": True,
+        "alpha_divergence": True,
+        # following values are >0
+        "beta_divergence": False,
+        "ab_divergence": False,
+        "renyi_divergence": False,
+        "l1_distance": False,
+        "l2_distance": False,
+        "l_infinity_distance": False,
+        "fisher_rao_distance": False,
+    }
 
     def __init__(
         self,
@@ -155,6 +168,15 @@ class InfoLM(Metric):
         self.add_state("preds_attention_mask", [], dist_reduce_fx="cat")
         self.add_state("target_input_ids", [], dist_reduce_fx="cat")
         self.add_state("target_attention_mask", [], dist_reduce_fx="cat")
+
+    @property
+    def higher_is_better(self) -> bool:  # type: ignore[override]
+        """Returns a bool indicating whether a higher value of the information measure is better.
+
+        Done this way as depends on if the information measure is positive or negative.
+
+        """
+        return self._information_measure_higher_is_better[self.information_measure]
 
     def update(self, preds: Union[str, Sequence[str]], target: Union[str, Sequence[str]]) -> None:
         """Update state with predictions and targets."""